ATO-130: Move logic relating to query parameters validation out of OrchestrationAuthorizationService and into new QueryParamsValidationService
